requirements.txt
Flask==3.1.2
gunicorn==21.2.0
requests==2.32.5
PyPDF2
python-docx
